[opt] 优化MsgProcessor；timeout类型改为double以支持更高精度的超时

+++ framework/amulet/src/main/java/com/kedacom/vconf/sdk/amulet/IMagicBook.java
@@ -63,7 +63,7 @@ public interface IMagicBook {
      * NOTE： 对于没有响应的请求此方法不适用
      * @return 超时时长，单位：秒。
      * */
-    int timeout(@NonNull String reqId);
+    double timeout(@NonNull String reqId);
 
     /**
      * 响应名称对应的响应ID集合
